requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
peewee==3.18.2
pillow==11.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.4
requests==2.32.5
six==1.17.0
sniffio==1.3.1
urllib3==2.5.0
